Debugging sloted BO
Time offset issues. We tried to simulate simultaneous events using random time ranges. However, "clock drifts" are an issue to consider.

+++ structures/notification.h
@@ -66,14 +66,14 @@ struct TxInfo
 	double tx_gain;				// Transmission gain [linear ratio]
 	double data_rate; 			// Rate at which data is transmitted
 	int *modulation_schemes;	// Modulation scheme used
-	int x;						// X position of source node
-	int y;						// Y position of source node
-	int z;						// Z position of source node
+	double x;						// X position of source node
+	double y;						// Y position of source node
+	double z;						// Z position of source node
 	double nav_time;			// RTS/CTS NAV time
 
 	void PrintTxInfo(void){
 		printf("packet_id = %d - destination_id = %d - tx_duration = %f - tx_power = %f pw"
-				" - position = (%d, %d, %d)\n",
+				" - position = (%.2f, %.2f, %.2f)\n",
 				packet_id, destination_id, tx_duration, tx_power, x, y, z);
 	}
